requirements/common.txt
PyYAML==5.3.1
hbmqtt==0.9.6
